requirements/local.txt
-r base.txt
pytest==7.2.2
pytest-django==4.5.2
factory-boy==3.2.1
django_debug_toolbar==3.8.1
